+++ src/ts/ir/index.ts
@@ -10,7 +10,7 @@ import {
     selectEvent,
 } from "../util/editorCommonEvent";
 import {paste} from "../util/fixBrowserBehavior";
-import {hasClosestByClassName} from "../util/hasClosest";
+import {hasClosestByAttribute, hasClosestByClassName} from "../util/hasClosest";
 import {
     getEditorRange, setRangeByWbr,
     setSelectionFocus,
@@ -94,7 +94,7 @@ class IR {
                 this.preventInput = false;
                 return;
             }
-            if (this.composingLock ||  event.data === "‘" || event.data === "“" || event.data === "《") {
+            if (this.composingLock || event.data === "‘" || event.data === "“" || event.data === "《") {
                 return;
             }
             input(vditor, getSelection().getRangeAt(0).cloneRange(), false, event);
@@ -140,6 +140,12 @@ class IR {
                     range.selectNode(linkElement);
                     setSelectionFocus(range);
                 }
+            }
+            // 打开链接
+            const aElement = hasClosestByAttribute(event.target, "data-type", "a");
+            if (aElement && (!aElement.classList.contains("vditor-ir__node--expand"))) {
+                window.open(aElement.querySelector(":scope > .vditor-ir__marker--link").textContent);
+                return;
             }
 
             if (event.target.isEqualNode(this.element) && this.element.lastElementChild && range.collapsed) {