+++ src/ts/wysiwyg/index.ts
@@ -412,6 +412,11 @@ class WYSIWYG {
                 return;
             }
 
+            // 打开链接
+            if (event.target.tagName === "A") {
+                window.open(event.target.getAttribute("href"));
+            }
+
             const range = getEditorRange(vditor);
             if (event.target.isEqualNode(this.element) && this.element.lastElementChild && range.collapsed) {
                 const lastRect = this.element.lastElementChild.getBoundingClientRect();