+++ CHANGELOG.md
@@ -98,6 +98,7 @@
 
 ### v3.8.2 / 2021-02-xx
 
+* [751](https://github.com/Vanessa219/vditor/issues/751) firefox 代码块中 enter 键换行问题 `修复缺陷`
 * [923](https://github.com/Vanessa219/vditor/issues/923) 隐藏预览操作栏 `改进功能`
 
 ### v3.8.1 / 2021-02-04