+++ CHANGELOG.md
@@ -98,7 +98,8 @@
 
 ### v3.8.3 / 2021-02-xx
 
-* [949](https://github.com/Vanessa219/vditor/issues/949) lute.Md2HTML 3.7.5 后解析无 id 属性 `修复缺陷` TODO
+* [946](https://github.com/Vanessa219/vditor/issues/946) Improve the delete behavior in the list item after the code block `修复缺陷`
+* [949](https://github.com/Vanessa219/vditor/issues/949) lute.Md2HTML 3.7.5 后解析无 id 属性 `修复缺陷`
 * [936](https://github.com/Vanessa219/vditor/issues/936) Auto link `.app` domain suffix `改进功能`
 * [926](https://github.com/Vanessa219/vditor/issues/926) 即时渲染和所见即所得模式支持点击链接 `引入特性`